pyproject.toml
[tool.poetry]
name = "poetry add fastapi"
version = "0.1.0"
description = ""
authors = ["Enrique Manjavacas <enrique.manjavacas@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
toml = "^0.10.2"
pydantic-settings = "^2.4.0"
rispy = "^0.9.0"
fastapi-pagination = "^0.12.26"
uvicorn = "^0.30.5"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
humanize = "^4.10.0"
tqdm = "^4.66.5"
qdrant-client = "^1.11.0"
motor = "^3.5.1"
aiofiles = "^24.1.0"
aiohttp = "^3.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
